+++ tibiaauto-pub/mods/mod_creatureinfo/Util.cpp
@@ -50,12 +50,12 @@ char *CUtil::wc2c(const XMLCh *src)
 char * CUtil::getNodeAttribute(DOMNode *node, char *attrName)
 {
 	
-	int attrNr;
-	for (attrNr=0;attrNr<node->getAttributes()->getLength();attrNr++)
+	size_t attrNr;
+	for (attrNr = 0; attrNr < node->getAttributes()->getLength(); attrNr++)
 	{
 		DOMNode *attrNode = node->getAttributes()->item(attrNr);
-		char *attrNameWide=CUtil::wc2c(attrNode->getNodeName());
-		if (!strcmp(attrNameWide,attrName))
+		char *attrNameWide = CUtil::wc2c(attrNode->getNodeName());
+		if (!strcmp(attrNameWide, attrName))
 		{
 			free(attrNameWide);
 			return CUtil::wc2c(attrNode->getNodeValue());
