+++ tibiaauto-pub/mods/mod_responder/Util.cpp
@@ -49,8 +49,7 @@ char *CUtil::wc2c(const unsigned short *src)
 
 char * CUtil::getNodeAttribute(DOMNode *node, char *attrName)
 {
-	
-	int attrNr;
+	size_t attrNr;
 	for (attrNr=0;attrNr<node->getAttributes()->getLength();attrNr++)
 	{
 		DOMNode *attrNode = node->getAttributes()->item(attrNr);
